\name{cite.geoRglm}
\alias{cite.geoRglm}

\title{Citing Package geoRglm in Publications}
\description{
  How to cite the package geoRglm in publications.
}
\usage{
cite.geoRglm()
}
%- maybe also 'usage' for other objects documented here.
\details{
  Execute function \code{cite.geoRglm()} for information on how to cite
  geoRglm in publications.
}



\seealso{\code{\link[geoR]{cite.geoR}} for how to cite the package
  geoR. }
\examples{
cite.geoRglm()
}
\keyword{misc}
