\name{geoRglm-internal}


\alias{.BC.inv}
\alias{.glm.krige.aux}
\alias{.krige.bayes.extnd}
\alias{.krige.conv.extnd}
\alias{.krige.glm.check.aux}
\alias{.mcmc.check.aux}
\alias{.model.glm.check.aux}
\alias{.model.glsm.mcmc.check.aux}
\alias{.multgauss}
\alias{.object.match.names}
\alias{.output.glm.check.aux}
\alias{.pmixed}
\alias{.pois.log.grf}
\alias{.pred.aux}
\alias{.pred.quan.aux}
\alias{.prior.glm.check.aux}



\title{geoRglm internal functions}

\description{
  These are functions internally called by other functions
  in the package \pkg{geoRglm} and not meant to be called by the user.
}


\author{
  Ole F. Christensen \email{olefc@birc.au.dk}, \cr
  Paulo J. Ribeiro Jr. \email{Paulo.Ribeiro@est.ufpr.br}.
}



\keyword{spatial}
\keyword{internal}


