\name{geoRglm-defunct}

\title{Defunct Functions in the Package geoRglm}

\alias{geoRglmdefunct}
\alias{pois.log.krige}

\usage{
geoRglmdefunct()
}

\description{
  The functions listed here are no longer part of the package \code{geoRglm} as they are not needed (any more).
}

\details{
  
  The following functions are now defunct:
  
  \enumerate{
    
    \item{\code{pois.log.krige}}{renamed to \code{\link{pois.krige}}.
      From \code{geoRglm_0.6-0}.  }
    
    
  }
  
  
}


\keyword{spatial}
\keyword{utilities}

